requirements.txt
# Core dependencies
selenium~=4.33.0
pandas~=2.2.3
numpy~=2.0.2
matplotlib~=3.9.4
seaborn~=0.13.2

# Web scraping
webdriver-manager~=4.0.2
selenium-stealth~=1.0.6
beautifulsoup4~=4.12.0

# Data processing
openpyxl~=3.1.0
python-dotenv~=1.0.0

# Telegram bot
python-telegram-bot~=20.7

# Development
pytest~=7.4.0
black~=23.0.0
flake8~=6.0.0